get posts with user information and sort by desc

+++ app/controllers/posts_controller.ts
@@ -1,9 +1,11 @@
 import { HttpContext } from '@adonisjs/core/http'
-import { createPostValidator, updatePostValidator } from '../validators/post.js'
+import {
+  createPostValidator,
+  updatePostValidator,
+  deletePostValidator,
+} from '../validators/post.js'
 import Post from '#models/post'
 import db from '@adonisjs/lucid/services/db'
-// import db from '@adonisjs/lucid/services/db'
-// import { Database } from '@adonisjs/lucid/database'
 
 export default class PostsController {
   public async login(ctx: HttpContext) {
@@ -29,21 +31,28 @@ export default class PostsController {
     }
   }
 
-  public async getLimitedPosts({ response, request }: HttpContext) {
+  public async getLimitedPostsByCategory({ response, request }: HttpContext) {
     try {
       // const posts = await Post.all()
       const limit = request.input('limit', 5)
       const page = request.input('page', 1)
-      const type = 'Technology'
+      const type = request.input('category', 'Travel')
 
-      const posts = await db
-        .from('posts')
-        .select(
-          'posts.*', // All columns from posts table
-          'post_categories.type as category_type' // Only 'type' column from post_categories
-        )
+      // const posts = await db
+      //   .from('posts')
+      //   .select(
+      //     'posts.*',
+      //     'post_categories.type as category_type'
+      //   )
+      //   .join('post_categories', 'posts.post_category_id', 'post_categories.post_category_id')
+      //   .where('post_categories.type', type)
+      //   .paginate(page, Number(limit))
+      const posts = await Post.query()
+        .preload('user')
+        .preload('postCategory')
         .join('post_categories', 'posts.post_category_id', 'post_categories.post_category_id')
         .where('post_categories.type', type)
+        .orderBy('postId', 'desc')
         .paginate(page, Number(limit))
 
       response.status(200).send(posts)
@@ -57,11 +66,26 @@ export default class PostsController {
 
   public async createPost({ request, response }: HttpContext) {
     try {
+      console.log(request.all())
       const validatedData = await createPostValidator.validate(request.all())
 
+      const category = await db
+        .from('post_categories')
+        .select('post_categories.*')
+        .where('post_categories.type', validatedData.category)
+
+      // check if category exists
+
+      if (category.length === 0) {
+        return response.status(404).send({
+          message: 'Category not found',
+        })
+      }
+
       const post = await Post.create({
-        userId: 1,
+        userId: validatedData.userId,
         text: validatedData.title,
+        postCategoryId: category[0].post_category_id,
       })
 
       return response.status(201).send({
@@ -78,7 +102,8 @@ export default class PostsController {
 
   public async updatePost({ request, response }: HttpContext) {
     const validatedData = await updatePostValidator.validate(request.all())
-    const postId: number = validatedData.id
+    const postId: number = validatedData.postId
+    const userId: number = validatedData.userId
     const newPostTitle: string = validatedData.title
 
     const post = await Post.findOrFail(postId)
@@ -89,7 +114,14 @@ export default class PostsController {
       })
     }
 
+    if (post.userId !== userId) {
+      return response.status(403).send({
+        message: 'You are not authorized to update this post',
+      })
+    }
+
     post.text = newPostTitle
+    await post.save()
     response.status(200).send({
       message: 'Update post successfully',
       post,
@@ -97,7 +129,9 @@ export default class PostsController {
   }
 
   public async deletePost({ request, response }: HttpContext) {
-    const postId: number = request.input('id')
+    const validatedData = await deletePostValidator.validate(request.all())
+    const postId: number = validatedData.postId
+    const userId: number = validatedData.userId
 
     const post = await Post.findOrFail(postId)
 
@@ -107,6 +141,11 @@ export default class PostsController {
       })
     }
 
+    if (post.userId !== userId) {
+      return response.status(403).send({
+        message: 'You are not authorized to delete this post',
+      })
+    }
     await post.delete()
 
     response.status(200).send({
